Move activity into a separate file
