+++ message_passing/introduction/__init__.py
@@ -1,0 +1,13 @@
+from enum import IntEnum
+
+TASK_QUEUE = "message-passing-introduction-task-queue"
+
+
+class Language(IntEnum):
+    ARABIC = 1
+    CHINESE = 2
+    ENGLISH = 3
+    FRENCH = 4
+    HINDI = 5
+    PORTUGUESE = 6
+    SPANISH = 7
