+++ message_passing/introduction/starter.py
@@ -3,14 +3,13 @@ from typing import Optional
 
 from temporalio.client import Client, WorkflowUpdateStage
 
+from message_passing.introduction import TASK_QUEUE
 from message_passing.introduction.workflows import (
     ApproveInput,
     GetLanguagesInput,
     GreetingWorkflow,
     Language,
 )
-
-TASK_QUEUE = "message-passing-introduction-task-queue"
 
 
 async def main(client: Optional[Client] = None):
